feat: add new plugin to theme mapstore with css file.
