+++ js/plugins.js
@@ -10,6 +10,7 @@ import productPlugins from '@mapstore/product/plugins';
 
 import NotAllowedPlugin from "@js/plugins/NotAllowed";
 import HeaderPlugin from "@js/plugins/Header";
+import EasyTheming from "@js/plugins/EasyTheming";
 import UserSessionPlugin from "@js/plugins/UserSession";
 import VersionPlugin from "@js/plugins/Version";
 import BackgroundSelectorPlugin from '@mapstore/plugins/BackgroundSelector';
@@ -30,6 +31,7 @@ export default {
         // georchestra plugins2
         NotAllowedPlugin,
         HeaderPlugin,
+        EasyTheming,
         UserSessionPlugin,
         VersionPlugin,
         BackgroundSelectorPlugin // This is not lazy loaded because of this issue. https://github.com/georchestra/mapstore2-georchestra/issues/573
